+++ src/lib/sketch/onboard/Onboard.svelte
@@ -111,8 +111,6 @@
 
 <style>
 	.onboard {
-		width: 100%;
-		height: 100%;
 		display: flex;
 		flex-direction: column;
 	}