+++ src/lib/sketch/onboard/consentful/Unburdensome.svelte
@@ -19,7 +19,7 @@
 			href="/help"
 			on:click|preventDefault={() =>
 				alert(
-					"We're still working on this, but if you actually want help you can email us at team@felt.social",
+					"Our real stuff is all in progress, but if you're interested in consentful design or Felt stuff, say hi at team@felt.social",
 				)}>felt.dev/help</a
 		>
 	</blockquote>