+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.8.1
 
+- add and publish `ui/Modal.svelte` and `ui/Portal.svelte`
+  ([#127](https://github.com/feltcoop/gro/pull/127))
 - add and publish `ui/PendingButton.svelte`
   ([#125](https://github.com/feltcoop/felt/pull/125))